requirements.txt
python-telegram-bot==20.7
tweepy==4.14.0
ollama==0.1.6
python-dotenv==1.0.0
requests==2.31.0
